Make required changes correctly
