+++ book.rb
@@ -1,0 +1,14 @@
+class Book
+  attr_accessor :title, :author
+  attr_reader :rentals
+
+  def initialize(title, author)
+    @title = title
+    @author = author
+    @rentals = []
+  end
+
+  def rent_book(person, date)
+    Rental.new(date, person, self)
+  end
+end
