+++ classroom.rb
@@ -12,31 +12,3 @@ class Classroom
     @students << student
   end
 end
-
-class Book
-  attr_accessor :title, :author
-  attr_reader :rentals
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-    @rentals = []
-  end
-
-  def add_rental(rental)
-    @rentals << rental
-  end
-end
-
-class Rental
-  attr_accessor :date
-  attr_reader :person, :book
-
-  def initialize(date, person, book)
-    @date = date
-    @person = person
-    @book = book
-    person.add_rental(self)
-    book.add_rental(self)
-  end
-end
