+++ main.rb
@@ -1,6 +1,9 @@
 require_relative 'person'
 require_relative 'student'
 require_relative 'teacher'
+require_relative 'classroom'
+require_relative 'book'
+require_relative 'rental'
 
 # Persons
