+++ person.rb
@@ -37,7 +37,7 @@ class Person < Nameable
     @name
   end
 
-  def add_rental(rental)
-    @rentals << rental
+  def add_book_rental(book, date)
+    Rental.new(date, book, self)
   end
 end
