+++ student.rb
@@ -4,8 +4,11 @@ require_relative 'classroom'
 class Student < Person
   attr_accessor :classroom
 
-  def initialize(age, classroom, name: 'Unknown', parent_permission: true)
+  def initialize(age, name: 'Unknown', parent_permission: true)
     super(age, name, parent_permission)
+  end
+
+  def assign_to_classroom(classroom)
     @classroom = classroom
     classroom.add_student(self)
   end
